+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Recommendation: for ease of reading, use the following order:
 - Fixed
 -->
 
+## [Unreleased]
+### Changed
+- Replaced default GraphQL playground with better maintained `graphiql` (old playground is still available)
+- Improved API server web console looks
+
 ## [0.56.1] - 2025-03-14
 ### Fixed
 - Value of flow agent throttling now presented in seconds